refactor: change urls location
