+++ components/navbar/navbar.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import dynamic from "next/dynamic";
 import Image from "next/image";
 import Link from "next/link";
-import urls from "../../urls/urls.json";
+import urls from "../../utils/urls.json";
 import { MenuItem } from "./components/menu-item/menu-item";
 import { MenuButton } from "./components/menu-button/menu-button";
